Update readme.txt to work better with some text editors
Some Text editors (such as Notepad++ and Sublime) allow you to click on a URL in a .txt file and then your default browser will navigate to it just fine. However, these same editors will sometimes get confused by a period at the end of a link and will add it, often leading to a 404.

This change simply removes the periods at the end of the two links in the text file allowing these readers to navigate properly.